refactor, and add a code to connect to the db
