+++ backend/src/app.ts
@@ -1,7 +1,8 @@
 import 'reflect-metadata';
 import { inject, injectable } from 'inversify';
 import { HttpServerService } from './services/HttpServerService';
-import { TYPES } from './types';
+import { TYPES } from './container/types';
+import { DatabaseService } from './services/DatabaseService';
 
 export type AppConfig = {
   port: number,
@@ -11,9 +12,11 @@ export type AppConfig = {
 export class App {
   constructor(
     @inject(TYPES.HttpServerService) private readonly http: HttpServerService,
+    @inject(TYPES.DatabaseService) private readonly db: DatabaseService,
   ) {}
 
   async run(): Promise<void> {
+    await this.db.connect();
     return this.http.listen();
   }
 }
