+++ backend/src/controllers/RootController.ts
@@ -1,4 +1,4 @@
-import { controller } from "inversify-express-utils";
+import { controller } from 'inversify-express-utils';
 
 @controller('/')
 export class RootController {
