+++ backend/src/services/AccountService.ts
@@ -1,6 +1,6 @@
 import { Container, inject, injectable } from 'inversify';
-import { TYPES } from '../../types';
-import { DatabaseService } from '../DatabaseService';
+import { TYPES } from '../container/types';
+import { DatabaseService } from './DatabaseService';
 
 @injectable()
 export class AccountService {
